# server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: KMJ Optical ERP API
  version: 0.1.0
  description: OpenAPI spec for core KMJ ERP endpoints (auth, customers, orders, bills, payments, inventory, delivery)
servers:
  - url: http://localhost:4000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/orders:
    get:
      summary: List orders
      responses:
        "200": { description: OK }
    post:
      summary: Create order
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                customerId: { type: string }
                frame: { type: string }
                lens: { type: string }
      responses:
        "200": { description: created }
  /api/orders/{id}:
    get:
      summary: Get order
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
...
security:
  - bearerAuth: []